// src/app/core/interfaces/user.interface.ts
export interface IUser{
  UID?: string;
  name?: string;
  email?: string;
  phone?:number;
  age?:number;
  city?:string;
  skills?:string;
  jobDesc?: string;
  availableTime?: string;
  courses?: string;
  type?:string;
  website?: string;
  logo?: string;
  password?: string;
  isUser?:boolean;
}